requirements.txt
google-adk[extensions]>=1.0.0
asyncpg>=0.29.0
minio>=7.2.0
openai>=1.0.0
pydantic-settings>=2.0.0
playwright>=1.44.0
mcp>=1.0.0
httpx>=0.27.0
